Extract suggestion generators into single-purpose modules
