+++ src/engine/assistant/__tests__/recursive-evo.test.ts
@@ -36,7 +36,7 @@ describe('Recursive Evolution Suggestions', () => {
       mockSaveData as unknown as import('../../saveParser/index').SaveData,
       false,
       null,
-      mockApiData as unknown as import('../suggestionEngine').AssistantApiData,
+      mockApiData as unknown as import('../strategies/types').AssistantApiData,
       mockStrategy,
     );
