+++ src/engine/assistant/__tests__/suggestionEngine.fix.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it } from 'vitest';
 import type { PokemonMetadata } from '../../../db/schema';
 import type { SaveData } from '../../saveParser/index';
 import { gen1Strategy } from '../strategies/gen1Strategy';
-import type { AssistantApiData } from '../suggestionEngine';
+import type { AssistantApiData } from '../strategies/types';
 import { generateSuggestions } from '../suggestionEngine';
 
 describe('suggestionEngine - Redundancy Fix Verification', () => {
