+++ src/engine/assistant/generators/index.ts
@@ -1,0 +1,4 @@
+export * from './breed';
+export * from './catch';
+export * from './evolution';
+export * from './tradeGift';
