+++ src/engine/assistant/generators/tradeGift.ts
@@ -1,0 +1,155 @@
+import type { StaticGift } from '../../data/gen1/assistantData';
+import { STATIC_NPC_TRADE_DATA } from '../../data/gen1/assistantData';
+import { getUnobtainableReason } from '../../exclusives/gen1Exclusives';
+import { getGen2UnobtainableReason } from '../../exclusives/gen2Exclusives';
+import type { PokemonInstance, SaveData } from '../../saveParser/index';
+import type { AssistantApiData, Suggestion } from '../strategies/types';
+
+/**
+ * Checks if a specific bit flag is set within a continuous byte array.
+ * Used primarily for validating whether in-game event flags (like catching a static encounter
+ * or claiming a gift Pokémon) have been triggered in the player's save file.
+ *
+ * @param flags - The raw byte array extracted from the save file representing a block of event flags.
+ * @param flagId - The specific zero-indexed bit ID to check.
+ * @returns True if the bit is set (1), false if it is unset (0) or if the inputs are invalid.
+ */
+export function checkFlag(flags: Uint8Array | undefined, flagId: number | undefined) {
+  if (!flags || flagId === undefined) return false;
+  const byteIndex = flagId >> 3;
+  const bitIndex = flagId & 7;
+  const byte = flags[byteIndex];
+  if (byte === undefined) return false;
+  return (byte & (1 << bitIndex)) !== 0;
+}
+
+/**
+ * Evaluates version exclusives, in-game NPC trades, and static gift encounters.
+ *
+ * It mutates the provided `suggestions` array.
+ *
+ * @param queryTargets - The top priority missing Pokémon IDs to evaluate.
+ * @param saveData - The player's parsed save file, containing badges and event flags.
+ * @param displayVersion - The current game version string.
+ * @param ownedSet - A Set of Pokémon IDs the player already owns.
+ * @param apiData - Pre-fetched metadata for Pokémon definitions.
+ * @param instancesBySpecies - A Map of the player's physical Pokémon, used to check for required trade offerings or pre-evolutions.
+ * @param suggestions - The shared array where new suggestions are pushed.
+ * @param missingIds - A Set of Pokémon IDs the player needs to obtain.
+ * @param staticGiftEntries - The pre-computed array of static gift entries from suggestionEngine.
+ */
+export function generateGiftAndTradeSuggestions(
+  queryTargets: number[],
+  saveData: SaveData,
+  displayVersion: string,
+  ownedSet: Set<number>,
+  apiData: AssistantApiData,
+  instancesBySpecies: Map<number, PokemonInstance[]>,
+  suggestions: Suggestion[],
+  missingIds: Set<number>,
+  staticGiftEntries: { giftId: number; gift: StaticGift }[],
+) {
+  // B. Unobtainable / Exclusive logic
+  // Checks if the target is completely locked out of the current version (e.g. Red exclusives on Blue).
+  // These are assigned the lowest base priority (10) since they require external action (link cable trades).
+  // ⚡ Bolt: Convert O(N^2) array lookup to O(1) Set has for NPC trades
+  const validNpcTradeIds = new Set<number>();
+  for (let i = 0; i < STATIC_NPC_TRADE_DATA.length; i++) {
+    const t = STATIC_NPC_TRADE_DATA[i];
+    if (t && t.gen === saveData.generation && (!t.versions || t.versions.includes(displayVersion))) {
+      validNpcTradeIds.add(t.receivedId);
+    }
+  }
+
+  const pidsWithExclusives = new Set<number>();
+  for (const pid of queryTargets) {
+    let reason: string | null = null;
+    if (saveData.generation === 2) {
+      reason = getGen2UnobtainableReason(pid, displayVersion, ownedSet.size, ownedSet);
+    } else {
+      reason = getUnobtainableReason(pid, displayVersion, ownedSet.size, ownedSet);
+    }
+    if (reason) {
+      pidsWithExclusives.add(pid);
+
+      const isNpcTrade = validNpcTradeIds.has(pid);
+      if (isNpcTrade) continue;
+
+      // If they physically own a pre-evolution, they don't strictly need to trade, they can evolve it!
+      const p = apiData.pokemonMetadata?.[pid];
+      let hasPhysicalPreEvo = false;
+      if (p?.efrm && p.efrm.length > 0) {
+        // Iterate backwards through all ancestors (recursive ownership check)
+        // The current logic only checked immediate parents, so if a player had Charmander,
+        // Charizard might incorrectly be flagged as Unobtainable/Trade.
+        for (let i = p.efrm.length - 1; i >= 0; i--) {
+          const preId = p.efrm[i];
+          if (preId !== undefined && instancesBySpecies.has(preId)) {
+            hasPhysicalPreEvo = true;
+            break;
+          }
+        }
+      }
+      if (hasPhysicalPreEvo) continue;
+
+      suggestions.push({
+        id: `exclusive-${pid}`,
+        category: 'Trade',
+        title: `Version Exclusive: #${pid}`,
+        description: reason,
+        pokemonId: pid,
+        priority: 10,
+      });
+    }
+  }
+
+  // C. In-Game NPC Trades
+  // Priority boosts if the player already physically possesses the required "offered" Pokemon (65 -> 85).
+  for (const trade of STATIC_NPC_TRADE_DATA) {
+    if (trade.gen !== saveData.generation) continue;
+    if (trade.versions && !trade.versions.includes(displayVersion)) continue;
+    if (!missingIds.has(trade.receivedId)) continue;
+
+    if (trade.tradeIndex !== undefined && saveData.npcTradeFlags !== undefined) {
+      const isClaimed = (saveData.npcTradeFlags & (1 << trade.tradeIndex)) !== 0;
+      if (isClaimed) continue;
+    }
+
+    const hasOffered = instancesBySpecies.has(trade.offeredId);
+    suggestions.push({
+      id: `npc-trade-${trade.receivedId}`,
+      category: 'Trade',
+      title: `Trade for #${trade.receivedId}`,
+      description: hasOffered
+        ? `You have #${trade.offeredId}! Trade it at ${trade.location} for #${trade.receivedId}.`
+        : `Catch #${trade.offeredId} and trade it at ${trade.location} for #${trade.receivedId}.`,
+      pokemonId: trade.receivedId,
+      priority: hasOffered ? 85 : 65,
+    });
+  }
+
+  // D. Static Gifts
+  // Suggests available static encounters and gifts that haven't been claimed yet.
+  for (let i = 0; i < staticGiftEntries.length; i++) {
+    const entry = staticGiftEntries[i];
+    if (!entry) continue;
+    const { giftId, gift } = entry;
+    if (gift.gen && gift.gen !== saveData.generation) continue;
+    if (!missingIds.has(giftId)) continue;
+
+    const requiredBadges = gift.requiredBadges || 0;
+    if (saveData.badges < requiredBadges) continue;
+
+    const hasClaimed = checkFlag(saveData.eventFlags, gift.eventFlag);
+    if (hasClaimed) continue;
+
+    suggestions.push({
+      id: `gift-${giftId}`,
+      category: 'Gift',
+      title: `Claim Gift: #${giftId}`,
+      description: `Get ${gift.name} at ${gift.location} (${gift.reason}).`,
+      pokemonId: giftId,
+      priority: 85,
+    });
+  }
+}
